get rid of warnings in interleave

+++ contrib/interleave_fastq.c
@@ -23,7 +23,7 @@ int main(int argc, char *argv[])
 
   if(  argc != 4  || !strcmp(argv[1], "-h") ||  !strcmp(argv[1], "-help") ) {
    printf("\nThe fast program to interleave two possibly gzipped FASTQ files into one\n");
-   printf("(Version 1.7)  Process the read length up to 4,096,000 \n"); 
+   printf("(Version 1.7)  Process the read length up to 4,096,000 \n");
    printf("Usage :  %s <Read1_Fastq_file_name>[.gz] <Read2_Fastq_file_name>[.gz] <Interleaved_Fastq_file_name>\n", argv[0]);
    printf("Input file: Read1_Fastq_file_name[.gz]\n");
    printf("Input file: Read2_Fastq_file_name[.gz]\n");
@@ -31,7 +31,7 @@ int main(int argc, char *argv[])
    if (argc <= 1 || !strcmp(argv[1], "-h") ||  !strcmp(argv[1], "-help")) exit(0);
    else exit(1);
   }
- 
+
   gzFile FQ1, FQ2;
   FILE *IFQ;
   kseq_t *kseq1, *kseq2;
@@ -63,7 +63,7 @@ int main(int argc, char *argv[])
   const char * empty = "";
   const char * p1 = "/1";
   const char * p2 = "/2";
-  while((len = kseq_read(kseq1)) > 0) 
+  while((len = kseq_read(kseq1)) > 0)
   {
     if ((len = kseq_read(kseq2)) <= 0) {
       fprintf(stderr, "Read 1 and Read 2 have different record counts!\n");
@@ -71,19 +71,19 @@ int main(int argc, char *argv[])
     }
     int is_same = 1;
     if (kseq1->name.l != kseq2->name.l) {
-      fprintf(stderr, "Read 1 and Read 2 have different read names %.*s %.*s\n",  kseq1->name.l, kseq1->name.s, kseq2->name.l, kseq2->name.s);
+      fprintf(stderr, "Read 1 and Read 2 have different read names %.*s %.*s\n",  (int)kseq1->name.l, kseq1->name.s, (int)kseq2->name.l, kseq2->name.s);
       exit(1);
     }
     if (memcmp(kseq1->name.s, kseq2->name.s, kseq1->name.l) != 0) {
       if (kseq1->name.l <= 2 || memcmp(kseq1->name.s, kseq2->name.s, kseq1->name.l - 2) != 0) {
-        fprintf(stderr, "Read 1 and Read 2 have are not paired read names %.*s %.*s\n",  kseq1->name.l, kseq1->name.s, kseq2->name.l, kseq2->name.s);
+        fprintf(stderr, "Read 1 and Read 2 have are not paired read names %.*s %.*s\n",  (int)kseq1->name.l, kseq1->name.s, (int)kseq2->name.l, kseq2->name.s);
         exit(1);
       }
       is_same = 0;
     }
-      
-    fprintf(IFQ, "@%.*s%s\n%.*s\n+\n%.*s\n", kseq1->name.l, kseq1->name.s, (is_same ? p1 : empty), kseq1->seq.l, kseq1->seq.s, kseq1->qual.l, kseq1->qual.s);
-    fprintf(IFQ, "@%.*s%s\n%.*s\n+\n%.*s\n", kseq2->name.l, kseq2->name.s, (is_same ? p2 : empty), kseq2->seq.l, kseq2->seq.s, kseq2->qual.l, kseq2->qual.s);
+
+    fprintf(IFQ, "@%.*s%s\n%.*s\n+\n%.*s\n", (int)kseq1->name.l, kseq1->name.s, (is_same ? p1 : empty), (int)kseq1->seq.l, kseq1->seq.s, (int)kseq1->qual.l, kseq1->qual.s);
+    fprintf(IFQ, "@%.*s%s\n%.*s\n+\n%.*s\n", (int)kseq2->name.l, kseq2->name.s, (is_same ? p2 : empty), (int)kseq2->seq.l, kseq2->seq.s, (int)kseq2->qual.l, kseq2->qual.s);
   }
 
   kseq_destroy(kseq1);
